Modify sliceable to use longs directly as opposed to Numerics.

+++ src/main/java/prj/clark/alchemy/data/AlchemyList.java
@@ -20,7 +20,7 @@ public class AlchemyList implements Sequenced, Sliceable, Printable {
     }
 
     @Override
-    public Sequenced slice(Numeric start, Numeric end, Numeric n) {
+    public Sequenced slice(long start, long end, long n) {
         // TODO(matthew-c21) - Implement using a lazy sequence.
         return null;
     }
